requirements.txt
certifi==2020.6.20
cffi==1.14.3
click==7.1.2
cryptography==3.2.1
Flask==1.1.2
Flask-Cors==3.0.9
PyJWT==1.7.1
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
python-dateutil==2.8.1
six==1.15.0
SQLAlchemy==1.3.20
Werkzeug==1.0.1
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
cffi==1.14.3
click==7.1.2
cryptography==3.2.1
Flask==1.1.2
Flask-Cors==3.0.9
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
python-dateutil==2.8.1
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
